move changes to separate subproject "inject-generics"
