+++ src/main/java/org/mockito/internal/configuration/plugins/DefaultMockitoPlugins.java
@@ -11,6 +11,8 @@ import java.util.Set;
 
 import org.mockito.MockMakers;
 import org.mockito.internal.util.MockUtil;
+import org.mockito.internal.configuration.injection.filter.MockCandidateFilter;
+import org.mockito.internal.configuration.injection.filter.TypeBasedCandidateFilter;
 import org.mockito.plugins.AnnotationEngine;
 import org.mockito.plugins.DoNotMockEnforcerWithType;
 import org.mockito.plugins.InstantiatorProvider2;
@@ -64,6 +66,7 @@ public class DefaultMockitoPlugins implements MockitoPlugins {
         DEFAULT_PLUGINS.put(
                 DoNotMockEnforcerWithType.class.getName(),
                 "org.mockito.internal.configuration.DefaultDoNotMockEnforcer");
+        DEFAULT_PLUGINS.put(MockCandidateFilter.class.getName(), TypeBasedCandidateFilter.class.getName());
 
         MOCK_MAKER_ALIASES.add(INLINE_ALIAS);
         MOCK_MAKER_ALIASES.add(PROXY_ALIAS);
