+++ src/main/java/org/mockito/internal/configuration/plugins/PluginRegistry.java
@@ -6,6 +6,7 @@ package org.mockito.internal.configuration.plugins;
 
 import java.util.List;
 
+import org.mockito.internal.configuration.injection.filter.MockCandidateFilter;
 import org.mockito.plugins.AnnotationEngine;
 import org.mockito.plugins.DoNotMockEnforcer;
 import org.mockito.plugins.DoNotMockEnforcerWithType;
@@ -51,6 +52,9 @@ class PluginRegistry {
     private final DoNotMockEnforcerWithType doNotMockEnforcer =
             new PluginLoader(pluginSwitch)
                     .loadPlugin(DoNotMockEnforcerWithType.class, DoNotMockEnforcer.class);
+
+    private final MockCandidateFilter mockCandidateFilter =
+            new PluginLoader(pluginSwitch).loadPlugin(MockCandidateFilter.class);
 
     PluginRegistry() {
         instantiatorProvider =
@@ -133,5 +137,14 @@ class PluginRegistry {
      */
     List<MockResolver> getMockResolvers() {
         return mockResolvers;
+    }
+
+    /**
+     * Returns the filter applied for handling @InjectMocks.
+     * @return {@link org.mockito.internal.configuration.injection.filter.TypeBasedCandidateFilter} if no
+     *         {@link MockCandidateFilter} (internal) extension exists or is visible in the current classpath.
+     */
+    MockCandidateFilter getMockCandidateFilter() {
+        return mockCandidateFilter;
     }
 }
