+++ src/test/java/org/mockitousage/plugins/MockitoPluginsTest.java
@@ -4,10 +4,14 @@
  */
 package org.mockitousage.plugins;
 
+import static org.junit.Assert.assertFalse;
 import static org.junit.Assert.assertNotNull;
 
+import java.lang.reflect.Proxy;
 import org.junit.Test;
 import org.mockito.Mockito;
+import org.mockito.internal.configuration.injection.filter.MockCandidateFilter;
+import org.mockito.internal.configuration.plugins.InternalPlugins;
 import org.mockito.plugins.AnnotationEngine;
 import org.mockito.plugins.InstantiatorProvider2;
 import org.mockito.plugins.MockMaker;
@@ -30,5 +34,13 @@ public class MockitoPluginsTest extends TestBase {
         assertNotNull(plugins.getDefaultPlugin(InstantiatorProvider2.class));
         assertNotNull(plugins.getDefaultPlugin(AnnotationEngine.class));
         assertNotNull(plugins.getDefaultPlugin(MockitoLogger.class));
+    }
+
+    @Test
+    public void provides_internal_plugins() {
+        MockCandidateFilter mockCandidateFilter = InternalPlugins.getMockCandidateFilter();
+        assertNotNull(mockCandidateFilter);
+        // ensure PluginLoader did not create a throwing proxy instance
+        assertFalse(Proxy.isProxyClass(mockCandidateFilter.getClass()));
     }
 }
