+++ subprojects/inject-generics/src/main/java/org/mockito/internal/util/reflection/generic/MatchArrayClass.java
