+++ subprojects/inject-generics/src/main/resources/mockito-extensions/org.mockito.internal.configuration.injection.filter.MockCandidateFilter
@@ -1,0 +1,1 @@
+org.mockito.internal.configuration.injection.filter.TypeWithGenericsCandidateFilter
