+++ subprojects/inject-generics/src/test/java/org/mockitousage/regression/Regression3000Test.java
