+++ subprojects/junit-jupiter/junit-jupiter.gradle
@@ -16,7 +16,6 @@ dependencies {
     implementation libraries.junitJupiterApi
     testImplementation libraries.assertj
     testImplementation libraries.junitPlatformLauncher
-    testImplementation libraries.junitJupiterParams
     testRuntimeOnly libraries.junitJupiterEngine
 }
